Move ClassLoaderSafe* clases to presto-plugin-toolkit
