+++ presto-phoenix/src/main/java/io/prestosql/plugin/phoenix/PhoenixConnectorFactory.java
@@ -19,6 +19,9 @@ import com.google.inject.util.Types;
 import io.airlift.bootstrap.Bootstrap;
 import io.airlift.bootstrap.LifeCycleManager;
 import io.airlift.json.JsonModule;
+import io.prestosql.plugin.base.classloader.ClassLoaderSafeConnectorMetadata;
+import io.prestosql.plugin.base.classloader.ClassLoaderSafeConnectorPageSinkProvider;
+import io.prestosql.plugin.base.classloader.ClassLoaderSafeConnectorSplitManager;
 import io.prestosql.plugin.base.jmx.MBeanServerModule;
 import io.prestosql.plugin.jdbc.SessionPropertiesProvider;
 import io.prestosql.spi.classloader.ThreadContextClassLoader;
@@ -29,9 +32,6 @@ import io.prestosql.spi.connector.ConnectorHandleResolver;
 import io.prestosql.spi.connector.ConnectorPageSinkProvider;
 import io.prestosql.spi.connector.ConnectorRecordSetProvider;
 import io.prestosql.spi.connector.ConnectorSplitManager;
-import io.prestosql.spi.connector.classloader.ClassLoaderSafeConnectorMetadata;
-import io.prestosql.spi.connector.classloader.ClassLoaderSafeConnectorPageSinkProvider;
-import io.prestosql.spi.connector.classloader.ClassLoaderSafeConnectorSplitManager;
 import org.weakref.jmx.guice.MBeanModule;
 
 import java.util.Map;
